Renames all namespaces and urls
