+++ test/TestableHttpClient.IntegrationTests/AssertingRequests.cs
@@ -1,10 +1,9 @@
 ﻿using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.IntegrationTests
+namespace TestableHttpClient.IntegrationTests
 {
     public class AssertingRequests
     {
