+++ test/TestableHttpClient.IntegrationTests/ConfigureResponses.cs
@@ -2,10 +2,9 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.IntegrationTests
+namespace TestableHttpClient.IntegrationTests
 {
     public class ConfigureResponses
     {
