+++ test/TestableHttpClient.Tests/HttpRequestMessageAsserterTests.cs
@@ -5,10 +5,9 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.Tests
+namespace TestableHttpClient.Tests
 {
     public class HttpRequestMessageAsserterTests
     {
