+++ test/TestableHttpClient.Tests/HttpRequestMessageExtensionsTests.HasContentHeader.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Net.Http;
 using System.Net.Http.Headers;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.Tests
+namespace TestableHttpClient.Tests
 {
     public partial class HttpRequestMessageExtensionsTests
     {
