+++ test/TestableHttpClient.Tests/HttpRequestMessageExtensionsTests.HasHttpVersion.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Net;
 using System.Net.Http;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.Tests
+namespace TestableHttpClient.Tests
 {
     public partial class HttpRequestMessageExtensionsTests
     {
