+++ test/TestableHttpClient.Tests/HttpResponseMessageBuilderTests.cs
@@ -4,10 +4,9 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.Tests
+namespace TestableHttpClient.Tests
 {
     public class HttpResponseMessageBuilderTests
     {
