+++ test/TestableHttpClient.Tests/TestableHttpMessageHandlerTests.cs
@@ -2,10 +2,9 @@
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-
 using Xunit;
 
-namespace HttpClientTestHelpers.Tests
+namespace TestableHttpClient.Tests
 {
     public class TestableHttpMessageHandlerTests
     {
